transformation in plater + gokuSS1

+++ src/main/java/com/andres_k/component/network/networkSend/messageServer/MessageStatePlayer.java
@@ -7,8 +7,8 @@ import com.andres_k.component.network.networkSend.MessageModel;
  */
 public class MessageStatePlayer extends MessageModel {
     private float life;
-    private int ki;
-    private int energy;
+    private float ki;
+    private float energy;
 
     public MessageStatePlayer() {
     }
@@ -17,11 +17,11 @@ public class MessageStatePlayer extends MessageModel {
         return this.life;
     }
 
-    public int getKi() {
+    public float getKi() {
         return this.ki;
     }
 
-    public int getEnergy() {
+    public float getEnergy() {
         return this.energy;
     }
 
@@ -29,11 +29,11 @@ public class MessageStatePlayer extends MessageModel {
         this.life = life;
     }
 
-    public void setKi(int ki) {
+    public void setKi(float ki) {
         this.ki = ki;
     }
 
-    public void setEnergy(int energy) {
+    public void setEnergy(float energy) {
         this.energy = energy;
     }
